backend/requirements.txt
# FastAPI and ASGI Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# AI/ML
openai==1.3.7

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# WebSocket
websockets==12.0

# Utilities
python-dotenv==1.0.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
